Moved MenuItems from CheckingScreen to NavigationMenu
